requirements.txt
africastalking==1.2.9
amqp==5.3.1
asgiref==3.8.1
attrs==25.1.0
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.10
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
msgpack==1.1.0
oauthlib==3.2.2
prompt_toolkit==3.0.50
psycopg==3.2.4
psycopg-binary==3.2.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
